Resolve Find Your Twin lint checks
